Refactor: Response 패키지 변경 (#5)
DTO 내 구분을 위해서 패키지를 새로 생성하여 분리하였다.
